+++ src/Assets/AssetContainer.php
@@ -3,6 +3,7 @@
 namespace Statamic\Assets;
 
 use ArrayAccess;
+use Illuminate\Contracts\Support\Arrayable;
 use Statamic\Contracts\Assets\AssetContainer as AssetContainerContract;
 use Statamic\Contracts\Data\Augmentable;
 use Statamic\Contracts\Data\Augmented;
@@ -21,7 +22,7 @@ use Statamic\Facades\Stache;
 use Statamic\Facades\URL;
 use Statamic\Support\Traits\FluentlyGetsAndSets;
 
-class AssetContainer implements AssetContainerContract, Augmentable, ArrayAccess
+class AssetContainer implements AssetContainerContract, Augmentable, ArrayAccess, Arrayable
 {
     use ExistsAsFile, FluentlyGetsAndSets, HasAugmentedInstance;
 
@@ -92,36 +93,14 @@ class AssetContainer implements AssetContainerContract, Augmentable, ArrayAccess
         return URL::makeAbsolute($this->url());
     }
 
-    /**
-     * Convert to an array.
-     *
-     * @return array
-     */
-    public function toArray()
-    {
-        $array = [
-            'title' => $this->title,
-            'handle' => $this->handle,
-            'disk' => $this->disk,
-            'search_index' => $this->searchIndex,
-            'allow_uploads' => $this->allowUploads,
-            'allow_downloading' => $this->allowDownloading,
-            'allow_renaming' => $this->allowRenaming,
-            'allow_moving' => $this->allowMoving,
-            'create_folders' => $this->createFolders,
-        ];
-
-        // if ($user = user()) {
-        //     $array['allow_uploads'] = user()->can('store', [AssetContract::class, $this]);
-        //     $array['create_folders'] = user()->can('create', [AssetFolder::class, $this]);
-        // }
-
-        return $array;
-    }
-
     public function newAugmentedInstance(): Augmented
     {
         return new AugmentedAssetContainer($this);
+    }
+
+    protected function excludedEvaluatedAugmentedArrayKeys()
+    {
+        return ['assets'];
     }
 
     /**
@@ -465,19 +444,16 @@ class AssetContainer implements AssetContainerContract, Augmentable, ArrayAccess
 
     public function fileData()
     {
-        $data = array_except($this->toArray(), 'handle');
-
-        // @TODO: Determine why we were explicity unsetting this data
-
-        // if ($data['allow_uploads'] === true) {
-        //     unset($data['allow_uploads']);
-        // }
-
-        // if ($data['create_folders'] === true) {
-        //     unset($data['create_folders']);
-        // }
-
-        return $data;
+        return [
+            'title' => $this->title,
+            'disk' => $this->disk,
+            'search_index' => $this->searchIndex,
+            'allow_uploads' => $this->allowUploads,
+            'allow_downloading' => $this->allowDownloading,
+            'allow_renaming' => $this->allowRenaming,
+            'allow_moving' => $this->allowMoving,
+            'create_folders' => $this->createFolders,
+        ];
     }
 
     public function toCacheableArray()