+++ src/Query/Builder.php
@@ -15,6 +15,7 @@ abstract class Builder implements Contract
     protected $limit;
     protected $offset = 0;
     protected $wheres = [];
+    protected $with = [];
     protected $orderBys = [];
     protected $operators = [
         '=' => 'Equals',
@@ -396,5 +397,12 @@ abstract class Builder implements Contract
     protected function filterTestNotLikeRegex($item, $pattern)
     {
         return ! $this->filterTestLikeRegex($item, $pattern);
+    }
+
+    public function with($relations)
+    {
+        $this->with = array_merge($this->with, Arr::wrap($relations));
+
+        return $this;
     }
 }