+++ src/Structures/Page.php
@@ -3,6 +3,7 @@
 namespace Statamic\Structures;
 
 use ArrayAccess;
+use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Contracts\Support\Responsable;
 use Illuminate\Support\Traits\ForwardsCalls;
 use JsonSerializable;
@@ -22,7 +23,7 @@ use Statamic\Facades\Site;
 use Statamic\Facades\URL;
 use Statamic\GraphQL\ResolvesValues;
 
-class Page implements Entry, Augmentable, Responsable, Protectable, JsonSerializable, ResolvesValuesContract, ArrayAccess
+class Page implements Entry, Augmentable, Responsable, Protectable, JsonSerializable, ResolvesValuesContract, ArrayAccess, Arrayable
 {
     use HasAugmentedInstance, ForwardsCalls, TracksQueriedColumns, ResolvesValues, ContainsSupplementalData;
 
@@ -431,6 +432,8 @@ class Page implements Entry, Augmentable, Responsable, Protectable, JsonSerializ
     #[\ReturnTypeWillChange]
     public function jsonSerialize()
     {
+        // hmmm? ->with() ?
+
         return $this
             ->toAugmentedCollection($this->selectedQueryColumns)
             ->withShallowNesting()