+++ tests/Data/Taxonomies/LocalizedTermTest.php
@@ -4,10 +4,14 @@ namespace Tests\Data\Taxonomies;
 
 use BadMethodCallException;
 use Facades\Statamic\Fields\BlueprintRepository;
+use Illuminate\Contracts\Support\Arrayable;
 use Mockery;
 use Statamic\Facades;
+use Statamic\Facades\Blueprint;
 use Statamic\Facades\Taxonomy;
 use Statamic\Fields\Fieldtype;
+use Statamic\Fields\Value;
+use Statamic\Support\Arr;
 use Statamic\Taxonomies\LocalizedTerm;
 use Statamic\Taxonomies\Term;
 use Tests\PreventSavingStacheItemsToDisk;
@@ -148,5 +152,110 @@ class LocalizedTermTest extends TestCase
 
         Taxonomy::make('tags')->save();
         (new Term)->taxonomy('tags')->in('en')->thisFieldDoesntExist();
+    }
+
+    /** @test */
+    public function it_converts_to_an_array()
+    {
+        $fieldtype = new class extends Fieldtype
+        {
+            protected static $handle = 'test';
+
+            public function augment($value)
+            {
+                return [
+                    new Value('alfa'),
+                    new Value([
+                        new Value('bravo'),
+                        new Value('charlie'),
+                        'delta',
+                    ]),
+                ];
+            }
+        };
+        $fieldtype::register();
+
+        $blueprint = Blueprint::makeFromFields([
+            'baz' => [
+                'type' => 'test',
+            ],
+        ]);
+        BlueprintRepository::shouldReceive('in')->with('taxonomies/tags')->andReturn(collect([
+            'tags' => $blueprint->setHandle('tags'),
+        ]));
+
+        Taxonomy::make('tags')->save();
+        $term = (new Term)->taxonomy('tags')->slug('test')->dataForLocale('en', [
+            'foo' => 'bar',
+            'baz' => 'qux',
+        ])->in('en');
+
+        $this->assertInstanceOf(Arrayable::class, $term);
+
+        $array = $term->toArray();
+        $this->assertEquals($term->augmented()->keys(), array_keys($array));
+        $this->assertEquals([
+            'alfa',
+            [
+                'bravo',
+                'charlie',
+                'delta',
+            ],
+        ], $array['baz'], 'Value objects are not resolved recursively');
+
+        $array = $term
+            ->selectedQueryColumns($keys = ['id', 'foo', 'baz'])
+            ->toArray();
+
+        $this->assertEquals($keys, array_keys($array), 'toArray keys differ from selectedQueryColumns');
+    }
+
+    /** @test */
+    public function only_requested_relationship_fields_are_included_in_to_array()
+    {
+        $regularFieldtype = new class extends Fieldtype
+        {
+            protected static $handle = 'regular';
+
+            public function augment($value)
+            {
+                return 'augmented '.$value;
+            }
+        };
+        $regularFieldtype::register();
+
+        $relationshipFieldtype = new class extends Fieldtype
+        {
+            protected static $handle = 'relationship';
+            protected $relationship = true;
+
+            public function augment($values)
+            {
+                return collect($values)->map(fn ($value) => 'augmented '.$value)->all();
+            }
+        };
+        $relationshipFieldtype::register();
+
+        $blueprint = Blueprint::makeFromFields([
+            'alfa' => ['type' => 'regular'],
+            'bravo' => ['type' => 'relationship'],
+            'charlie' => ['type' => 'relationship'],
+        ]);
+        BlueprintRepository::shouldReceive('in')->with('taxonomies/tags')->andReturn(collect([
+            'tags' => $blueprint->setHandle('tags'),
+        ]));
+
+        Taxonomy::make('tags')->save();
+        $term = (new Term)->taxonomy('tags')->slug('test')->dataForLocale('en', [
+            'alfa' => 'one',
+            'bravo' => ['a', 'b'],
+            'charlie' => ['c', 'd'],
+        ])->in('en');
+
+        $this->assertEquals([
+            'alfa' => 'augmented one',
+            'bravo' => ['a', 'b'],
+            'charlie' => ['augmented c', 'augmented d'],
+        ], Arr::only($term->selectedQueryRelations(['charlie'])->toArray(), ['alfa', 'bravo', 'charlie']));
     }
 }